--- theme-contract/src/main/java/com/orange/ouds/theme/components/OudsComponentsTokens.kt
@@ -1,17 +1,0 @@
-/*
- * Software Name: OUDS Android
- * SPDX-FileCopyrightText: Copyright (c) Orange SA
- * SPDX-License-Identifier: MIT
- *
- * This software is distributed under the MIT license,
- * the text of which is available at https://opensource.org/license/MIT/
- * or see the "LICENSE" file for more details.
- *
- * Software description: Android library of reusable graphical components
- */
-
-package com.orange.ouds.theme.components
-
-class OudsComponentsTokens(
-    val button: OudsButtonTokens = OudsButtonTokens(),
-)
